Fix `using: true` schema dump issue for Rails 4.0.x
When using v2.x of this gem you are required to use a particular version
of rgeo that *is not* compatible with this. Specifically, if you are on
Rails version 4.0.x you will see your spatial indices dumped into your
schema file as `using: true`. That's because the postgis adapter is trying
to use a true/false identifier for whether or not an index is spatial and
passes that to rgeo, except rgeo is expecting a raw value it can embed in
the options hash, so when a true value is given it outputs `using: true`.
This appears to be solved in later versions of Rails since this gem and
rgeo both seem to "catch up" to each other in future versions that are
spec'd beyond Rails 4.0.x.